Add program item validity check also to backend
